Make CDArray element counts 64-bit-safe (pointer hygiene, stage 3g)
CDArray::CommitToOffset/DecommitAboveOffset set the committed-element
count from a pointer difference ((T*)pv - atArray), and CFastHeap::Reset
passes (char*)p_reset_position - atArray as a length. On x64 those
differences are 64-bit ptrdiff_t implicitly narrowed to the uint count,
which is a C4244 warning (the values are real element counts that fit, so
this is a warning, not a truncation bug). Make the narrowing explicit
with (uint).

These are the narrowings deferred from the FastHeap slice (3825cd3); they
live in the widely-instantiated CDArray<T> template, so keeping them out
of that commit kept it focused on the actual address-truncation fix.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings; the 18 C4291 citing FastHeap remain the pre-existing
placement-new-without-delete baseline), smoke-launched OK. Seventh slice
of the uint32->uintptr sweep after VirtualMem (ba41823), FastHeap
(3825cd3), loaders (43c4b87), audio (1d560a9), Mesh (0491ee3), MemoryLog
(fbf5098).

+++ jp2_pc/Source/Lib/Sys/FastHeap.hpp
@@ -515,7 +515,7 @@ private:
 			return;
 
 		void* pv = pvCommitMem(atArray + uCommit, (u_offset - uCommit) * sizeof(T));
-		uCommit = (T*)pv - atArray;
+		uCommit = (uint)((T*)pv - atArray);
 	}
 
 	//*****************************************************************************************
@@ -535,7 +535,7 @@ private:
 			return;
 
 		void* pv = pvDecommitMem(atArray + u_offset, (uCommit+1 - u_offset) * sizeof(T) - 1);
-		uCommit = (T*)pv - atArray;
+		uCommit = (uint)((T*)pv - atArray);
 	}
 
 };
@@ -698,7 +698,7 @@ public:
 		Assert(bWithin((char*)p_reset_position, atArray, atArray + uLen));
 		Assert((uintptr)p_reset_position % uHEAP_ALIGNMENT == 0);
 
-		CDArray<char>::Reset((char*)p_reset_position - atArray,false,u4_decommit);
+		CDArray<char>::Reset((uint)((char*)p_reset_position - atArray),false,u4_decommit);
 	}
